Rename `mars_rover_interface.py` -> `rover_api.py`
